+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/GetSpecialFollowingsRequest.cs
@@ -13,7 +13,7 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos
             Mid = userId;
         }
 
-        public GetSpecialFollowingsRequest(long userId, int tagId)
+        public GetSpecialFollowingsRequest(long userId, long tagId)
         {
             Mid = userId;
             Tagid = tagId;
@@ -25,7 +25,7 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos
         /// TagId
         /// </summary>
         /// <sample>-10:特别关注</sample>
-        public int Tagid { get; set; } = -10;
+        public long Tagid { get; set; } = -10;
 
         public int Pn { get; set; } = 1;