+++ src/Ray.BiliBiliTool.Agent/BiliBiliAgent/Dtos/Live/AreaDto.cs
@@ -14,7 +14,7 @@ namespace Ray.BiliBiliTool.Agent.BiliBiliAgent.Dtos.Live
 
     public class AreaDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
     }