+++ CHANGELOG.md
@@ -102,3 +102,8 @@
 - Featur: Replace AOP from MethodBoundaryAspect.Fody to Rougamo.Fody, to fix async exception
 - Merge PR #448 : Fix typo
 - Fix #446 : Change id type from int to long
+## 1.0.1
+- Fix #463 : do not trust user's ck config
+- Feature #460 : publish single file when release
+- Feature: use new scripts for gh actions's release
+- Feature #473 : let user input when there is no target task in configs